Move battery to its own file
